feat : 포스트와 포스트 태그 엔티티 수정 및 서비스 로직 리팩토링

+++ backend/src/main/java/wooteco/prolog/post/application/dto/PostsResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Getter
 public class PostsResponse {
     private List<PostResponse> data;
-    private int totalSize;
+    private Long totalSize;
     private int totalPage;
     private int currPage;
 }
